Add manpages.
We adapt the manpages from the Fedora package, which were written by Jerry
James using text from the sources.  The binary names have been changed
to reflect the topcom- prefix we use in the Debian packages, and we
also substitute the upstream version number at build instead of having
it hardcoded.